+++ bungee/src/main/java/com/github/games647/fastlogin/bungee/FastLoginBungee.java
@@ -50,6 +50,7 @@ import java.util.List;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ThreadFactory;
 
+import net.md_5.bungee.BungeeServerInfo;
 import net.md_5.bungee.api.CommandSender;
 import net.md_5.bungee.api.chat.TextComponent;
 import net.md_5.bungee.api.connection.PendingConnection;
@@ -188,5 +189,10 @@ public class FastLoginBungee extends Plugin implements PlatformPlugin<CommandSen
     @Override
     public AsyncScheduler getScheduler() {
         return scheduler;
+    }
+
+    @Override
+    public boolean isPluginInstalled(String name) {
+        return getProxy().getPluginManager().getPlugin(name) != null;
     }
 }
